requirements.txt
numpy==2.2.5
Shimmy==2.0.0
gymnasium==1.1.1
dm_control==1.0.28
mujoco==3.3.0
imageio==2.37.0
